Move some functions to a utility file
